refactor: isolate credential pool administration methods
